% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{castDATEToTime}
\alias{castDATEToTime}
\title{Convert a dataset's columns from DATE string format to numeric}
\usage{
castDATEToTime(dataset, colnames)
}
\arguments{
\item{dataset}{the dataset you want to convert}

\item{colnames}{the names of the columns you want to convert}
}
\description{
Convert a dataset's columns from DATE string format to numeric
}
\seealso{
Other Data formatting: \code{\link{castDATETIMEToTime}},
  \code{\link{castStringToNum}},
  \code{\link{castTIMEToTime}}, \code{\link{setFieldNames}}
}
