% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{castStringToNum}
\alias{castStringToNum}
\title{Convert a dataset's columns from string to numeric}
\usage{
castStringToNum(dataset, colnames)
}
\arguments{
\item{dataset}{the dataset you want to convert}

\item{colnames}{the names of the columns you want to convert}
}
\description{
Convert a dataset's columns from string to numeric
}
\seealso{
Other Data formatting: \code{\link{castDATETIMEToTime}},
  \code{\link{castDATEToTime}},
  \code{\link{castTIMEToTime}}
}
