% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{castTIMEToTime}
\alias{castTIMEToTime}
\title{Convert a dataset's columns from TIME string format to numeric}
\usage{
castTIMEToTime(dataset, colnames)
}
\arguments{
\item{dataset}{the dataset you want to convert}

\item{colnames}{the names of the columns you want to convert}
}
\description{
Convert a dataset's columns from TIME string format to numeric
}
\seealso{
Other Data formatting: \code{\link{castDATETIMEToTime}},
  \code{\link{castDATEToTime}},
  \code{\link{castStringToNum}}
}
