% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{cleanDataset}
\alias{cleanDataset}
\title{Clean a dataset. Remove all rows containing missing values}
\usage{
cleanDataset(dataset)
}
\arguments{
\item{dataset}{the dataset you want to clean}
}
\description{
Clean a dataset. Remove all rows containing missing values
}
\seealso{
Other Data preparation: \code{\link{splitDataset}}
}
