% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{deleteFlexioReccord}
\alias{deleteFlexioReccord}
\title{Deletes a ressource reccord from Flexio}
\usage{
deleteFlexioReccord(flexioURL, account, ressourceName, reccordID, auth,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{reccordID}{: the ID of the reccord you want}

\item{auth}{flexio authentification token}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Deletes a ressource reccord from Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{getFlexioReccord}},
  \code{\link{getFlexioRessource}},
  \code{\link{postFlexioRessource}}
}
