% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{deleteFlexioRecord}
\alias{deleteFlexioRecord}
\title{Deletes a resource record from Flexio}
\usage{
deleteFlexioRecord(flexioURL, account, resourceName, recordID, auth,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{resourceName}{name of the flexio resource}

\item{recordID}{: the ID of the record you want}

\item{auth}{Flexio authentification token}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Deletes a resource record from Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{downloadFlexioFile}},
  \code{\link{getFlexioRecord}},
  \code{\link{getFlexioResourceFieldsTypes}},
  \code{\link{getFlexioResourceSchema}},
  \code{\link{getFlexioResource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioResource}},
  \code{\link{putFlexioRecord}}
}
