% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\docType{data}
\name{flexioPaginationLength}
\alias{flexioPaginationLength}
\title{Max range authorized by Flexio API}
\format{An object of class \code{numeric} of length 1.}
\usage{
flexioPaginationLength
}
\description{
Max range authorized by Flexio API
}
\keyword{datasets}
