\name{flexioR-package}
\alias{flexioR-package}
\alias{flexioR}
\docType{package}
\title{
\packageTitle{flexioR}
}
\description{
\packageDescription{flexioR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{flexioR}
\packageIndices{flexioR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{flexioR}

Maintainer: \packageMaintainer{flexioR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
