% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioReccord}
\alias{getFlexioReccord}
\title{Returns a 1 entry dataset containing a single ressource reccord from Flexio}
\usage{
getFlexioReccord(flexioURL, account, ressourceName, auth, reccordID,
  fields = c(), verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}

\item{reccordID}{: the ID of the reccord you want}

\item{fields}{name of the fields you want to get in your dataset (leave it empty if you want all the fields)}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Returns a 1 entry dataset containing a single ressource reccord from Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioReccord}},
  \code{\link{getFlexioRessource}},
  \code{\link{patchFlexioReccord}},
  \code{\link{postFlexioRessource}},
  \code{\link{putFlexioReccord}}
}
