% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioRecord}
\alias{getFlexioRecord}
\title{Returns a 1 entry dataset containing a single ressource record from Flexio}
\usage{
getFlexioRecord(flexioURL, account, ressourceName, auth, recordID,
  fields = c(), verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}

\item{recordID}{: the ID of the record you want}

\item{fields}{name of the fields you want to get in your dataset (leave it empty if you want all the fields)}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Returns a 1 entry dataset containing a single ressource record from Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{getFlexioRessourceFieldsTypes}},
  \code{\link{getFlexioRessourceSchema}},
  \code{\link{getFlexioRessource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioRessource}},
  \code{\link{putFlexioRecord}}
}
