% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioResource}
\alias{getFlexioResource}
\title{Returns a dataset containing resources records from Flexio}
\usage{
getFlexioResource(flexioURL, account, resourceName, auth, header = NULL,
  fields = c(), verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{resourceName}{name of the flexio resource}

\item{auth}{flexio authentification token}

\item{header}{additional header (you don't need to add 'range' and 'Authorization', they have been put automaticaly for you)}

\item{fields}{name of the fields you want to get in your dataset (leave it empty if you want all the fields)}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Returns a dataset containing resources records from Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{downloadFlexioFile}},
  \code{\link{getFlexioRecord}},
  \code{\link{getFlexioResourceFieldsTypes}},
  \code{\link{getFlexioResourceSchema}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioResource}},
  \code{\link{putFlexioRecord}}
}
