% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioResourceFieldsTypes}
\alias{getFlexioResourceFieldsTypes}
\title{Returns a list containing the Flexio resource fields names and their types}
\usage{
getFlexioResourceFieldsTypes(flexioURL, account, resourceName, auth)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{resourceName}{name of the flexio resource}

\item{auth}{flexio authentification token}
}
\description{
Returns a list containing the Flexio resource fields names and their types
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{getFlexioRecord}},
  \code{\link{getFlexioResourceSchema}},
  \code{\link{getFlexioResource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioResource}},
  \code{\link{putFlexioRecord}}
}
