% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioResourceSchema}
\alias{getFlexioResourceSchema}
\title{Returns the raw schema of a flexio resource}
\usage{
getFlexioResourceSchema(flexioURL, account, resourceName, auth,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{resourceName}{name of the flexio resource}

\item{auth}{flexio authentification token}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Returns the raw schema of a flexio resource
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{downloadFlexioFile}},
  \code{\link{getFlexioRecord}},
  \code{\link{getFlexioResourceFieldsTypes}},
  \code{\link{getFlexioResource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioResource}},
  \code{\link{putFlexioRecord}}
}
