% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioRessource}
\alias{getFlexioRessource}
\title{Gets a ressource from Flexio}
\usage{
getFlexioRessource(flexioURL, account, ressourceName, auth, header = NULL,
  fields = c(), verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}
}
\description{
Gets a ressource from Flexio
}
