% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioRessourceFieldsTypes}
\alias{getFlexioRessourceFieldsTypes}
\title{Returns a list containing the Flexio ressource fields names and their types}
\usage{
getFlexioRessourceFieldsTypes(flexioURL, account, ressourceName, auth)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}
}
\description{
Returns a list containing the Flexio ressource fields names and their types
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioReccord}},
  \code{\link{getFlexioReccord}},
  \code{\link{getFlexioRessourceSchema}},
  \code{\link{getFlexioRessource}},
  \code{\link{patchFlexioReccord}},
  \code{\link{postFlexioRessource}},
  \code{\link{putFlexioReccord}}
}
