% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{patchFlexioReccord}
\alias{patchFlexioReccord}
\title{Updates a ressource reccord from Flexio (only the given fields)}
\usage{
patchFlexioReccord(flexioURL, account, ressourceName, auth, reccordID, data,
  fields = c(), verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}

\item{reccordID}{: the ID of the reccord you want}

\item{data}{: the data you want to use}

\item{fields}{: the fields you want to update (leave it empty to update all the fields)}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Updates a ressource reccord from Flexio (only the given fields)
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioReccord}},
  \code{\link{getFlexioReccord}},
  \code{\link{getFlexioRessource}},
  \code{\link{postFlexioRessource}},
  \code{\link{putFlexioReccord}}
}
