% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{postFlexioResource}
\alias{postFlexioResource}
\title{Sends a resource to Flexio. Adds each entry of the given dataset to the Flexio resource, Returns the a vector with the recordIDs}
\usage{
postFlexioResource(flexioURL, account, resourceName, auth, data,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{resourceName}{name of the flexio resource}

\item{auth}{flexio authentification token}

\item{data}{the data you want to send to Flexio}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Sends a resource to Flexio. Adds each entry of the given dataset to the Flexio resource, Returns the a vector with the recordIDs
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{getFlexioRecord}},
  \code{\link{getFlexioResourceFieldsTypes}},
  \code{\link{getFlexioResourceSchema}},
  \code{\link{getFlexioResource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{putFlexioRecord}}
}
