% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{postFlexioRessource}
\alias{postFlexioRessource}
\title{Sends a ressource to Flexio}
\usage{
postFlexioRessource(flexioURL, account, ressourceName, auth, data,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}

\item{data}{the data you want to send to Flexio}

\item{verbose}{set it to TRUE to print the request details}

\item{fields}{name of the fields you want to get in your dataset (leave it empty if you want all the fields)}
}
\description{
Sends a ressource to Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioReccord}},
  \code{\link{getFlexioReccord}},
  \code{\link{getFlexioRessource}}
}
