% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{putFlexioReccord}
\alias{putFlexioReccord}
\title{Updates a ressource reccord from Flexio (all the fields)}
\usage{
putFlexioReccord(flexioURL, account, ressourceName, auth, reccordID, data,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}

\item{reccordID}{: the ID of the reccord you want}

\item{data}{: the data you want to use}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Updates a ressource reccord from Flexio (all the fields)
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioReccord}},
  \code{\link{getFlexioReccord}},
  \code{\link{getFlexioRessource}},
  \code{\link{patchFlexioReccord}},
  \code{\link{postFlexioRessource}}
}
