% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{returnValues}
\alias{returnValues}
\title{Writes a list as JSON to a file
If run inside Flexio, the data written in the file will then be used to by the connector}
\usage{
returnValues(valueList)
}
\arguments{
\item{valueList}{a list of named values (eg : "list(a:42, b:c(1,2,3))")}
}
\description{
Writes a list as JSON to a file
If run inside Flexio, the data written in the file will then be used to by the connector
}
