% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{saveDatasetToFile}
\alias{saveDatasetToFile}
\title{Saves a dataset in a .csv file}
\usage{
saveDatasetToFile(dataset, file = "tmp.csv")
}
\arguments{
\item{dataset}{the dataset you want to save}

\item{file}{the file you want to save the dataset in. Leave if empty to use the default file (tmp.csv)}
}
\description{
Saves a dataset in a .csv file
}
