% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{setFieldNames}
\alias{setFieldNames}
\title{Replaces the fields names with the given ones}
\usage{
setFieldNames(dataset, names)
}
\arguments{
\item{dataset}{the dataset you want to update names}

\item{names}{list of the old and new names formatted like that : list('old_name1'='new_name1', 'old_name2'='new_name2', ...)}
}
\description{
Replaces the fields names with the given ones
}
\seealso{
Other Data formatting: \code{\link{castDATETIMEToTime}},
  \code{\link{castDATEToTime}},
  \code{\link{castStringToNum}},
  \code{\link{castTIMEToTime}}
}
