% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{splitDataset}
\alias{splitDataset}
\title{Genrerates a training and a validation dataset with a given separation}
\usage{
splitDataset(dataset, separation)
}
\arguments{
\item{dataset}{the dataset you want to split}

\item{separation}{the repartition you want. Must be between 0 and 1}
}
\description{
Genrerates a training and a validation dataset with a given separation
}
\seealso{
Other Data preparation: \code{\link{cleanDataset}}
}
