% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{deleteFlexioRecord}
\alias{deleteFlexioRecord}
\title{Deletes a ressource record from Flexio}
\usage{
deleteFlexioRecord(flexioURL, account, ressourceName, recordID, auth,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{recordID}{: the ID of the record you want}

\item{auth}{flexio authentification token}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Deletes a ressource record from Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{getFlexioRecord}},
  \code{\link{getFlexioRessourceFieldsTypes}},
  \code{\link{getFlexioRessourceSchema}},
  \code{\link{getFlexioRessource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioRessource}},
  \code{\link{putFlexioRecord}}
}
