% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioRecord}
\alias{getFlexioRecord}
\title{Returns a 1 entry dataset containing a single resource record from Flexio}
\usage{
getFlexioRecord(flexioURL, account, resourceName, auth, recordID,
  fields = c(), verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{resourceName}{name of the flexio resource}

\item{auth}{flexio authentification token}

\item{recordID}{: the ID of the record you want}

\item{fields}{name of the fields you want to get in your dataset (leave it empty if you want all the fields)}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Returns a 1 entry dataset containing a single resource record from Flexio
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{downloadFlexioFile}},
  \code{\link{getFlexioResourceFieldsTypes}},
  \code{\link{getFlexioResourceSchema}},
  \code{\link{getFlexioResource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioResource}},
  \code{\link{putFlexioRecord}}
}
