% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{getFlexioRessourceSchema}
\alias{getFlexioRessourceSchema}
\title{Returns the raw schema of a flexio ressource}
\usage{
getFlexioRessourceSchema(flexioURL, account, ressourceName, auth,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Returns the raw schema of a flexio ressource
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{getFlexioRecord}},
  \code{\link{getFlexioRessourceFieldsTypes}},
  \code{\link{getFlexioRessource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioRessource}},
  \code{\link{putFlexioRecord}}
}
