% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{postFlexioRessource}
\alias{postFlexioRessource}
\title{Sends a ressource to Flexio. Adds each entry of the given dataset to the Flexio ressource}
\usage{
postFlexioRessource(flexioURL, account, ressourceName, auth, data,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{ressourceName}{name of the flexio ressource}

\item{auth}{flexio authentification token}

\item{data}{the data you want to send to Flexio}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Sends a ressource to Flexio. Adds each entry of the given dataset to the Flexio ressource
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioReccord}},
  \code{\link{getFlexioReccord}},
  \code{\link{getFlexioRessourceFieldsTypes}},
  \code{\link{getFlexioRessourceSchema}},
  \code{\link{getFlexioRessource}},
  \code{\link{patchFlexioReccord}},
  \code{\link{putFlexioReccord}}
}
