% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexio.r
\name{putFlexioRecord}
\alias{putFlexioRecord}
\title{Updates a resource record from Flexio (all the fields)}
\usage{
putFlexioRecord(flexioURL, account, resourceName, auth, recordID, data,
  verbose = FALSE)
}
\arguments{
\item{flexioURL}{URL of Flexio's API}

\item{account}{flexio account}

\item{resourceName}{name of the flexio resource}

\item{auth}{flexio authentification token}

\item{recordID}{: the ID of the record you want}

\item{data}{: the data you want to use}

\item{verbose}{set it to TRUE to print the request details}
}
\description{
Updates a resource record from Flexio (all the fields)
}
\seealso{
Other Flexio Interaction: \code{\link{deleteFlexioRecord}},
  \code{\link{getFlexioRecord}},
  \code{\link{getFlexioResourceFieldsTypes}},
  \code{\link{getFlexioResourceSchema}},
  \code{\link{getFlexioResource}},
  \code{\link{patchFlexioRecord}},
  \code{\link{postFlexioResource}}
}
